new scaffold

+++ PMIS/PMIS.Domain/Entities/EMPLOYEE_INFO.cs
@@ -10,11 +10,11 @@ namespace PMIS.Domain.Entities
             USER_INFOs = new HashSet<USER_INFO>();
         }
 
-        public decimal EMPLOYEE_ID { get; set; }
+        public int EMPLOYEE_ID { get; set; }
         public string? EMPLOYEE_CODE { get; set; }
         public string EMPLOYEE_NAME { get; set; } = null!;
         public string? EMPLOYEE_STATUS { get; set; }
-        public decimal? COMPANY_ID { get; set; }
+        public int? COMPANY_ID { get; set; }
 
         public virtual ICollection<USER_INFO> USER_INFOs { get; set; }
     }
