+++ PMIS/PMIS.Domain/Entities/NOTIFICATION_VIEW_POLICY.cs
@@ -5,12 +5,12 @@ namespace PMIS.Domain.Entities
 {
     public partial class NOTIFICATION_VIEW_POLICY
     {
-        public decimal ID { get; set; }
-        public decimal? NOTIFICATION_POLICY_ID { get; set; }
+        public int ID { get; set; }
+        public int? NOTIFICATION_POLICY_ID { get; set; }
         public string? STATUS { get; set; }
-        public decimal? COMPANY_ID { get; set; }
-        public decimal? UNIT_ID { get; set; }
-        public decimal? USER_ID { get; set; }
+        public int? COMPANY_ID { get; set; }
+        public int? UNIT_ID { get; set; }
+        public int? USER_ID { get; set; }
         public string? VIEW_PERMISSION { get; set; }
     }
 }
