+++ PMIS/PMIS.Domain/Entities/REPORT_CONFIGURATION.cs
@@ -7,9 +7,9 @@ namespace PMIS.Domain.Entities
     {
         public int REPORT_ID { get; set; }
         public string? REPORT_NAME { get; set; }
-        public decimal? MENU_ID { get; set; }
-        public decimal? ORDER_BY_SLNO { get; set; }
-        public decimal? COMPANY_ID { get; set; }
+        public int? MENU_ID { get; set; }
+        public int? ORDER_BY_SLNO { get; set; }
+        public int? COMPANY_ID { get; set; }
         public string? STATUS { get; set; }
         public string? ENTERED_BY { get; set; }
         public DateTime? ENTERED_DATE { get; set; }
