+++ PMIS/PMIS.Domain/Entities/RETURN_CAUSE_INFO.cs
@@ -5,7 +5,7 @@ namespace PMIS.Domain.Entities
 {
     public partial class RETURN_CAUSE_INFO
     {
-        public decimal RETURN_CAUSE_ID { get; set; }
+        public int RETURN_CAUSE_ID { get; set; }
         public string? RETURN_CAUSE_CODE { get; set; }
         public string? RETURN_CAUSE_NAME { get; set; }
         public string? STATUS { get; set; }
