+++ PMIS/PMIS.Domain/Entities/USER_DEFAULT_PAGE.cs
@@ -7,7 +7,7 @@ namespace PMIS.Domain.Entities
     {
         public int ID { get; set; }
         public int? MENU_ID { get; set; }
-        public decimal? USER_ID { get; set; }
+        public int? USER_ID { get; set; }
         public int COMPANY_ID { get; set; }
         public string? ENTERED_BY { get; set; }
         public DateTime? ENTERED_DATE { get; set; }
